refactor([tenant]/page): extract CatalogGrid client component
Pure move of search wrapper, chips, result-count line, and grid into a
new "use client" component. No behavior change — wires up to the same
chip-filtered visible array as before. Sets up Task 3 (real search input)
and Task 4 (filter logic) without touching layout or styling.

Spec deviation: chips render inside CatalogGrid (not page.tsx) because
the three regions must be sibling flex children of one Fragment.
<Link> URL-driven semantics are unchanged.
